Organize systematics and plot code into subdirectories
