+++ include/faint/plot/ChiSquaredCalculator.h
