+++ include/faint/plot/ErrorBandBuilder.h
