+++ include/faint/plot/HistogramExtremaCalculator.h
