+++ include/faint/plot/MatrixHistogramBuilder.h
