+++ include/faint/plot/PlotStyle.h
