+++ include/faint/plot/StackedHistogram.h
